--- a/src/hotspot/share/opto/c2compiler.cpp
+++ b/src/hotspot/share/opto/c2compiler.cpp
@@ -542,8 +542,14 @@ bool C2Compiler::is_intrinsic_supported(const methodHandle& method, bool is_virt
   case vmIntrinsics::_floatIsInfinite:
     if (!Matcher::match_rule_supported(Op_IsInfiniteF)) return false;
     break;
+  case vmIntrinsics::_floatIsFinite:
+    if (!Matcher::match_rule_supported(Op_IsFiniteF)) return false;
+    break;
   case vmIntrinsics::_doubleIsInfinite:
     if (!Matcher::match_rule_supported(Op_IsInfiniteD)) return false;
+    break;
+  case vmIntrinsics::_doubleIsFinite:
+    if (!Matcher::match_rule_supported(Op_IsFiniteD)) return false;
     break;
   case vmIntrinsics::_hashCode:
   case vmIntrinsics::_identityHashCode: