--- a/src/java.base/share/classes/java/lang/Double.java
+++ b/src/java.base/share/classes/java/lang/Double.java
@@ -782,6 +782,7 @@ public final class Double extends Number
      * floating-point value, {@code false} otherwise.
      * @since 1.8
      */
+    @IntrinsicCandidate
     public static boolean isFinite(double d) {
         return Math.abs(d) <= Double.MAX_VALUE;
     }