--- a/src/java.base/share/classes/java/lang/Float.java
+++ b/src/java.base/share/classes/java/lang/Float.java
@@ -604,6 +604,7 @@ public final class Float extends Number
      * floating-point value, {@code false} otherwise.
      * @since 1.8
      */
+     @IntrinsicCandidate
      public static boolean isFinite(float f) {
         return Math.abs(f) <= Float.MAX_VALUE;
     }