--- a/test/hotspot/jtreg/compiler/intrinsics/TestDoubleClassCheck.java
+++ b/test/hotspot/jtreg/compiler/intrinsics/TestDoubleClassCheck.java
@@ -21,16 +21,7 @@
  * questions.
  */
 
-/**
-* @test
-* @summary Test intrinsics for Double methods isNaN, isFinite, isInfinite.
-* @requires vm.cpu.features ~= ".*avx512dq.*" | os.arch == "riscv64"
-* @library /test/lib /
-* @run driver compiler.intrinsics.TestDoubleClassCheck
-*/
-
 package compiler.intrinsics;
-import compiler.lib.ir_framework.*;
 import java.util.random.RandomGenerator;
 import java.util.random.RandomGeneratorFactory;
 
@@ -39,10 +30,6 @@ public class TestDoubleClassCheck {
     int BUFFER_SIZE = 1024;
     double[] inputs;
     boolean[] outputs;
-
-    public static void main(String args[]) {
-        TestFramework.run(TestDoubleClassCheck.class);
-    }
 
     public TestDoubleClassCheck() {
         outputs = new boolean[BUFFER_SIZE];
@@ -58,17 +45,6 @@ public class TestDoubleClassCheck {
             inputs[i] = input;
         }
     }
-
-    @Test // needs to be run in (fast) debug mode
-    @Warmup(10000)
-    @IR(counts = {"IsInfiniteD", ">= 1"}) // Atleast one IsInfiniteD node is generated if intrinsic is used
-    public void testIsInfinite() {
-        for (int i = 0; i < BUFFER_SIZE; i++) {
-            outputs[i] = Double.isInfinite(inputs[i]);
-        }
-        checkResult("isInfinite");
-    }
-
 
     public void checkResult(String method) {
         for (int i=0; i < BUFFER_SIZE; i++) {