--- a/test/micro/org/openjdk/bench/java/lang/DoubleClassCheck.java
+++ b/test/micro/org/openjdk/bench/java/lang/DoubleClassCheck.java
@@ -102,4 +102,29 @@ public class DoubleClassCheck {
         }
     }
 
+    @Benchmark
+    @OperationsPerInvocation(BUFFER_SIZE)
+    public void testIsFiniteStore() {
+        for (int i = 0; i < BUFFER_SIZE; i++) {
+            storeOutputs[i] = Double.isFinite(inputs[i]);
+        }
+    }
+
+
+    @Benchmark
+    @OperationsPerInvocation(BUFFER_SIZE)
+    public void testIsFiniteCMov() {
+        for (int i = 0; i < BUFFER_SIZE; i++) {
+            cmovOutputs[i] = Double.isFinite(inputs[i]) ? 9 : 7;
+        }
+    }
+
+
+    @Benchmark
+    @OperationsPerInvocation(BUFFER_SIZE)
+    public void testIsFiniteBranch() {
+        for (int i = 0; i < BUFFER_SIZE; i++) {
+            cmovOutputs[i] = Double.isFinite(inputs[i]) ? call() : 7;
+        }
+    }
 }