… src/hotspot/share/opto/classes.hpp