Unlike on x86 (see 8285868 and the discussion in review), isFinite intrinsic turned out to be profitable on RISC-V using the same fclass instruction as for 8293695 (isInfinite instrinsic). Therefore, I'm proposing to have it added on RISC-V in this PR.

benchmark results:

before:

Benchmark                              Mode  Cnt   Score   Error  Units
DoubleClassCheck.testIsFiniteBranch    avgt   15  52.824 ± 1.744  ns/op
DoubleClassCheck.testIsFiniteCMov      avgt   15  16.104 ± 0.358  ns/op
DoubleClassCheck.testIsFiniteStore     avgt   15  14.366 ± 2.174  ns/op
FloatClassCheck.testIsFiniteBranch     avgt   15  49.821 ± 0.330  ns/op
FloatClassCheck.testIsFiniteCMov       avgt   15  14.702 ± 0.335  ns/op
FloatClassCheck.testIsFiniteStore      avgt   15  14.749 ± 0.496  ns/op

after:

DoubleClassCheck.testIsFiniteBranch    avgt   15  48.921 ± 0.557  ns/op
DoubleClassCheck.testIsFiniteCMov      avgt   15  13.716 ± 0.304  ns/op
DoubleClassCheck.testIsFiniteStore     avgt   15   9.152 ± 0.158  ns/op
FloatClassCheck.testIsFiniteBranch     avgt   15  47.740 ± 2.028  ns/op
FloatClassCheck.testIsFiniteCMov       avgt   15  13.299 ± 0.282  ns/op
FloatClassCheck.testIsFiniteStore      avgt   15   9.185 ± 0.396  ns/op

Existing isInfinite jtreg test was altered to be able to use common code for isFinite test and fine-grained requires tag filtering. Existing benchmark was modified to include isFinite case. A typo ("Atleast" -> "At least") was fixed on the way.

Test passed on both release and fastdebug builds. Hotspot tier1 tests were run on x86_64 and RISC-V with no issues.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8294198](https://bugs.openjdk.org/browse/JDK-8294198): Implement isFinite intrinsic for RISC-V


### Reviewers
 * [Fei Yang](https://openjdk.org/census#fyang) (@RealFYang - **Reviewer**)
 * @AryanP45 (no known github.com user name / role)
 * [Vladimir Kozlov](https://openjdk.org/census#kvn) (@vnkozlov - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk pull/10391/head:pull/10391` \
`$ git checkout pull/10391`

Update a local copy of the PR: \
`$ git checkout pull/10391` \
`$ git pull https://git.openjdk.org/jdk pull/10391/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 10391`

View PR using the GUI difftool: \
`$ git pr show -t 10391`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/10391.diff">https://git.openjdk.org/jdk/pull/10391.diff</a>

</details>