@RealFYang 2022-09-23
May I ask on which platform was the JMH test performed? I see some fluctuations on SiFive Unmatched board:

Before:
Benchmark                              Mode  Cnt   Score   Error  Units
DoubleClassCheck.testIsFiniteBranch    avgt   15  32.114 ± 3.514  ns/op
DoubleClassCheck.testIsFiniteCMov      avgt   15   8.656 ± 0.023  ns/op
DoubleClassCheck.testIsFiniteStore     avgt   15   7.757 ± 1.691  ns/op
FloatClassCheck.testIsFiniteBranch     avgt   15  32.446 ± 3.130  ns/op
FloatClassCheck.testIsFiniteCMov       avgt   15   7.801 ± 0.011  ns/op
FloatClassCheck.testIsFiniteStore      avgt   15   7.279 ± 1.483  ns/op

Benchmark                              Mode  Cnt   Score   Error  Units
DoubleClassCheck.testIsFiniteBranch    avgt   15  33.724 ± 5.157  ns/op
DoubleClassCheck.testIsFiniteCMov      avgt   15   8.642 ± 0.012  ns/op
DoubleClassCheck.testIsFiniteStore     avgt   15   7.765 ± 1.711  ns/op
FloatClassCheck.testIsFiniteBranch     avgt   15  31.401 ± 4.104  ns/op
FloatClassCheck.testIsFiniteCMov       avgt   15   7.799 ± 0.014  ns/op
FloatClassCheck.testIsFiniteStore      avgt   15   8.234 ± 0.023  ns/op

Benchmark                              Mode  Cnt   Score   Error  Units
DoubleClassCheck.testIsFiniteBranch    avgt   15  32.461 ± 3.688  ns/op
DoubleClassCheck.testIsFiniteCMov      avgt   15   8.643 ± 0.007  ns/op
DoubleClassCheck.testIsFiniteStore     avgt   15   7.772 ± 1.701  ns/op
FloatClassCheck.testIsFiniteBranch     avgt   15  33.258 ± 3.131  ns/op
FloatClassCheck.testIsFiniteCMov       avgt   15   7.815 ± 0.014  ns/op
FloatClassCheck.testIsFiniteStore      avgt   15   6.343 ± 1.475  ns/op

After:
Benchmark                              Mode  Cnt   Score   Error  Units
DoubleClassCheck.testIsFiniteBranch    avgt   15  33.915 ± 4.193  ns/op
DoubleClassCheck.testIsFiniteCMov      avgt   15   7.736 ± 0.012  ns/op
DoubleClassCheck.testIsFiniteStore     avgt   15   5.931 ± 0.007  ns/op
FloatClassCheck.testIsFiniteBranch     avgt   15  33.845 ± 3.449  ns/op
FloatClassCheck.testIsFiniteCMov       avgt   15   7.543 ± 0.013  ns/op
FloatClassCheck.testIsFiniteStore      avgt   15   6.035 ± 0.006  ns/op

Benchmark                              Mode  Cnt   Score   Error  Units
DoubleClassCheck.testIsFiniteBranch    avgt   15  33.421 ± 4.262  ns/op
DoubleClassCheck.testIsFiniteCMov      avgt   15   7.734 ± 0.012  ns/op
DoubleClassCheck.testIsFiniteStore     avgt   15   5.940 ± 0.010  ns/op
FloatClassCheck.testIsFiniteBranch     avgt   15  33.348 ± 2.656  ns/op
FloatClassCheck.testIsFiniteCMov       avgt   15   7.542 ± 0.013  ns/op
FloatClassCheck.testIsFiniteStore      avgt   15   6.040 ± 0.004  ns/op

Benchmark                              Mode  Cnt   Score   Error  Units
DoubleClassCheck.testIsFiniteBranch    avgt   15  31.669 ± 2.517  ns/op
DoubleClassCheck.testIsFiniteCMov      avgt   15   7.761 ± 0.092  ns/op
DoubleClassCheck.testIsFiniteStore     avgt   15   5.940 ± 0.007  ns/op
FloatClassCheck.testIsFiniteBranch     avgt   15  33.508 ± 3.680  ns/op
FloatClassCheck.testIsFiniteCMov       avgt   15   7.534 ± 0.009  ns/op
FloatClassCheck.testIsFiniteStore      avgt   15   6.031 ± 0.007  ns/op