@voitylov 2022-09-27
@vnkozlov could you or someone familiar with C2 look at shared C2 HotSpot changes? I did test it on x86, but any extra testing could help as well.