@voitylov 2022-09-29
Thank you @jddarcy @RealFYang @vnkozlov for your reviews.

I also ran the benchmarks with -XX:LoopUnrollLimit=0 -XX:LoopMaxUnroll=0 and see the improvement with these flags as well.